Finished viz for q8-11
cleaned data: the data from last year use "agree" etc. directly instead of mapping them to numbers --> to use the code from last year i map the numbers back to the strings

finished visualization for q8-11: since many of the questions include several sub-questions, I created files for each sub-questions separately.

To make it more convenient for myself to check, I allocated different local ports for different questions (see the end of the file)

For q10, I used QUESTION_ID to replace the hard-coded "question 10" so that it might be also useful for other similar questions in the future.